Move NoteNode to `shared-react`
