+++ packages/platform/src/editor/Editor.tsx
@@ -7,10 +7,10 @@ import { HistoryPlugin } from "@lexical/react/LexicalHistoryPlugin";
 import LexicalErrorBoundary from "@lexical/react/LexicalErrorBoundary";
 import { Usj } from "shared/converters/usj/usj.model";
 import scriptureUsjNodes from "shared/nodes/scripture/usj";
+import { NoteNode } from "shared-react/nodes/scripture/usj/NoteNode";
 import { LoggerBasic } from "shared-react/plugins/logger-basic.model";
 import UpdateStatePlugin from "shared-react/plugins/UpdateStatePlugin";
 import usjEditorAdaptor, { UsjNodeOptions } from "./adaptors/usj-editor.adaptor";
-import { NoteNode } from "./nodes/NoteNode";
 import editorTheme from "./themes/editor-theme";
 import ScriptureReferencePlugin from "./plugins/ScriptureReferencePlugin";
 import ToolbarPlugin from "./plugins/toolbar/ToolbarPlugin";
