+++ packages/shared-react/nodes/scripture/usj/NoteNode.tsx
